move SMTP sender access tests to separate file
