+++ notifications/notifications/src/test/kotlin/org/opensearch/integtest/SecurityNotificationIT.kt
@@ -507,838 +507,4 @@ class SecurityNotificationIT : PluginRestTestCase() {
 
         deleteUserWithCustomRole(user, NOTIFICATION_NO_ACCESS_ROLE)
     }
-
-    fun `test create smtp sender config with user that has create Notification permission`() {
-        createUserWithCustomRole(user, password, NOTIFICATION_CREATE_CONFIG_ACCESS, arrayOf(""), ROLE_TO_PERMISSION_MAPPING[NOTIFICATION_CREATE_CONFIG_ACCESS])
-
-        // Create sample config request reference
-        val sampleSmtpAccount = SmtpAccount("example-host", 2465, MethodType.SSL, "no-reply@fake-host.com")
-        val referenceObject = NotificationConfig(
-            "this is a sample config name",
-            "this is a sample config description",
-            ConfigType.SMTP_ACCOUNT,
-            isEnabled = true,
-            configData = sampleSmtpAccount
-        )
-
-        val sampleSmtpJsonString = getJsonString(sampleSmtpAccount)
-
-        // Create SMTP account config
-        val createRequestJsonString = """
-        {
-            "config":{
-                "name":"${referenceObject.name}",
-                "description":"${referenceObject.description}",
-                "config_type":"smtp_account",
-                "is_enabled":${referenceObject.isEnabled},
-                "smtp_account":$sampleSmtpJsonString
-            }
-        }
-        """.trimIndent()
-        try {
-            val configId = createConfigWithRequestJsonString(createRequestJsonString, userClient!!)
-            Assert.assertNotNull(configId)
-            Thread.sleep(1000)
-
-            // Get SMTP account config
-            val getConfigResponse = executeRequest(
-                RestRequest.Method.GET.name,
-                "${NotificationPlugin.PLUGIN_BASE_URI}/configs/$configId",
-                "",
-                RestStatus.OK.status
-            )
-            verifySingleConfigEquals(configId, referenceObject, getConfigResponse)
-        } finally {
-            deleteUserWithCustomRole(user, NOTIFICATION_CREATE_CONFIG_ACCESS)
-        }
-    }
-
-    fun `test get smtp sender has access with filter by backend roles enabled`() {
-        updateClusterSettings(ClusterSetting("persistent", PluginSettings.FILTER_BY_BACKEND_ROLES.key, true))
-
-        createUserWithCustomRole(user, password, NOTIFICATION_CREATE_CONFIG_ACCESS, arrayOf("role1"), ROLE_TO_PERMISSION_MAPPING[NOTIFICATION_CREATE_CONFIG_ACCESS])
-
-        // Create sample config request reference
-        val sampleSmtpAccount = SmtpAccount("example-host", 2465, MethodType.SSL, "no-reply@fake-host.com")
-        val referenceObject = NotificationConfig(
-            "this is a sample config name",
-            "this is a sample config description",
-            ConfigType.SMTP_ACCOUNT,
-            isEnabled = true,
-            configData = sampleSmtpAccount
-        )
-
-        val sampleSmtpJsonString = getJsonString(sampleSmtpAccount)
-
-        // Create SMTP account config
-        val createRequestJsonString = """
-        {
-            "config":{
-                "name":"${referenceObject.name}",
-                "description":"${referenceObject.description}",
-                "config_type":"smtp_account",
-                "is_enabled":${referenceObject.isEnabled},
-                "smtp_account":$sampleSmtpJsonString
-            }
-        }
-        """.trimIndent()
-
-        val getUser = "getUser"
-        val getUserClient = SecureRestClientBuilder(clusterHosts.toTypedArray(), isHttps(), getUser, password)
-            .setSocketTimeout(60000)
-            .setConnectionRequestTimeout(180000)
-            .build()
-
-        try {
-            val configId = createConfigWithRequestJsonString(createRequestJsonString, userClient!!)
-            Assert.assertNotNull(configId)
-            Thread.sleep(1000)
-
-            createUserWithCustomRole(getUser, password, NOTIFICATION_GET_CONFIG_ACCESS, arrayOf("role1"), ROLE_TO_PERMISSION_MAPPING[NOTIFICATION_GET_CONFIG_ACCESS])
-
-            // Get SMTP account config
-            val getConfigResponse = executeRequest(
-                RestRequest.Method.GET.name,
-                "${NotificationPlugin.PLUGIN_BASE_URI}/configs/$configId",
-                "",
-                RestStatus.OK.status,
-                getUserClient
-            )
-            verifySingleConfigEquals(configId, referenceObject, getConfigResponse)
-        } finally {
-            deleteUserWithCustomRole(user, NOTIFICATION_CREATE_CONFIG_ACCESS)
-            deleteUserWithCustomRole(getUser, NOTIFICATION_GET_CONFIG_ACCESS)
-            getUserClient?.close()
-        }
-    }
-
-    fun `test get smtp sender does not have access with filter by backend roles enabled`() {
-        updateClusterSettings(ClusterSetting("persistent", PluginSettings.FILTER_BY_BACKEND_ROLES.key, true))
-
-        createUserWithCustomRole(user, password, NOTIFICATION_CREATE_CONFIG_ACCESS, arrayOf("role1"), ROLE_TO_PERMISSION_MAPPING[NOTIFICATION_CREATE_CONFIG_ACCESS])
-
-        // Create sample config request reference
-        val sampleSmtpAccount = SmtpAccount("example-host", 2465, MethodType.SSL, "no-reply@fake-host.com")
-        val referenceObject = NotificationConfig(
-            "this is a sample config name",
-            "this is a sample config description",
-            ConfigType.SMTP_ACCOUNT,
-            isEnabled = true,
-            configData = sampleSmtpAccount
-        )
-
-        val sampleSmtpJsonString = getJsonString(sampleSmtpAccount)
-
-        // Create SMTP account config
-        val createRequestJsonString = """
-        {
-            "config":{
-                "name":"${referenceObject.name}",
-                "description":"${referenceObject.description}",
-                "config_type":"smtp_account",
-                "is_enabled":${referenceObject.isEnabled},
-                "smtp_account":$sampleSmtpJsonString
-            }
-        }
-        """.trimIndent()
-
-        val getUser = "getUser"
-        val getUserClient = SecureRestClientBuilder(clusterHosts.toTypedArray(), isHttps(), getUser, password)
-            .setSocketTimeout(60000)
-            .setConnectionRequestTimeout(180000)
-            .build()
-
-        try {
-            val configId = createConfigWithRequestJsonString(createRequestJsonString, userClient!!)
-            Assert.assertNotNull(configId)
-            Thread.sleep(1000)
-
-            createUserWithCustomRole(getUser, password, NOTIFICATION_GET_CONFIG_ACCESS, arrayOf("role2"), ROLE_TO_PERMISSION_MAPPING[NOTIFICATION_GET_CONFIG_ACCESS])
-
-            // Get SMTP account config
-            val getConfigResponse = executeRequest(
-                RestRequest.Method.GET.name,
-                "${NotificationPlugin.PLUGIN_BASE_URI}/configs/$configId",
-                "",
-                RestStatus.FORBIDDEN.status,
-                getUserClient
-            )
-        } finally {
-            deleteUserWithCustomRole(user, NOTIFICATION_CREATE_CONFIG_ACCESS)
-            deleteUserWithCustomRole(getUser, NOTIFICATION_GET_CONFIG_ACCESS)
-            getUserClient?.close()
-        }
-    }
-
-    fun `test get smtp sender has access when filter by backend access strategy is exact`() {
-        updateClusterSettings(ClusterSetting("persistent", PluginSettings.FILTER_BY_BACKEND_ROLES.key, true))
-        updateClusterSettings(ClusterSetting("persistent", PluginSettings.FILTER_BY_BACKEND_ROLES_ACCESS_STRATEGY.key, FilterByBackendRolesAccessStrategy.EXACT.strategy))
-
-        createUserWithCustomRole(user, password, NOTIFICATION_CREATE_CONFIG_ACCESS, arrayOf("role1", "role2"), ROLE_TO_PERMISSION_MAPPING[NOTIFICATION_CREATE_CONFIG_ACCESS])
-
-        // Create sample config request reference
-        val sampleSmtpAccount = SmtpAccount("example-host", 2465, MethodType.SSL, "no-reply@fake-host.com")
-        val referenceObject = NotificationConfig(
-            "this is a sample config name",
-            "this is a sample config description",
-            ConfigType.SMTP_ACCOUNT,
-            isEnabled = true,
-            configData = sampleSmtpAccount
-        )
-
-        val sampleSmtpJsonString = getJsonString(sampleSmtpAccount)
-
-        // Create SMTP account config
-        val createRequestJsonString = """
-        {
-            "config":{
-                "name":"${referenceObject.name}",
-                "description":"${referenceObject.description}",
-                "config_type":"smtp_account",
-                "is_enabled":${referenceObject.isEnabled},
-                "smtp_account":$sampleSmtpJsonString
-            }
-        }
-        """.trimIndent()
-
-        val getUser = "getUser"
-        val getUserClient = SecureRestClientBuilder(clusterHosts.toTypedArray(), isHttps(), getUser, password)
-            .setSocketTimeout(60000)
-            .setConnectionRequestTimeout(180000)
-            .build()
-
-        try {
-            val configId = createConfigWithRequestJsonString(createRequestJsonString, userClient!!)
-            Assert.assertNotNull(configId)
-            Thread.sleep(1000)
-
-            createUserWithCustomRole(getUser, password, NOTIFICATION_GET_CONFIG_ACCESS, arrayOf("role1", "role2"), ROLE_TO_PERMISSION_MAPPING[NOTIFICATION_GET_CONFIG_ACCESS])
-
-            // Get SMTP account config
-            val getConfigResponse = executeRequest(
-                RestRequest.Method.GET.name,
-                "${NotificationPlugin.PLUGIN_BASE_URI}/configs/$configId",
-                "",
-                RestStatus.OK.status,
-                getUserClient
-            )
-        } finally {
-            deleteUserWithCustomRole(user, NOTIFICATION_CREATE_CONFIG_ACCESS)
-            deleteUserWithCustomRole(getUser, NOTIFICATION_GET_CONFIG_ACCESS)
-            getUserClient?.close()
-        }
-    }
-
-    fun `test get smtp sender has access when filter by backend access strategy is all`() {
-        updateClusterSettings(ClusterSetting("persistent", PluginSettings.FILTER_BY_BACKEND_ROLES.key, true))
-        updateClusterSettings(ClusterSetting("persistent", PluginSettings.FILTER_BY_BACKEND_ROLES_ACCESS_STRATEGY.key, FilterByBackendRolesAccessStrategy.ALL.strategy))
-
-        createUserWithCustomRole(user, password, NOTIFICATION_CREATE_CONFIG_ACCESS, arrayOf("role1", "role2"), ROLE_TO_PERMISSION_MAPPING[NOTIFICATION_CREATE_CONFIG_ACCESS])
-
-        // Create sample config request reference
-        val sampleSmtpAccount = SmtpAccount("example-host", 2465, MethodType.SSL, "no-reply@fake-host.com")
-        val referenceObject = NotificationConfig(
-            "this is a sample config name",
-            "this is a sample config description",
-            ConfigType.SMTP_ACCOUNT,
-            isEnabled = true,
-            configData = sampleSmtpAccount
-        )
-
-        val sampleSmtpJsonString = getJsonString(sampleSmtpAccount)
-
-        // Create SMTP account config
-        val createRequestJsonString = """
-        {
-            "config":{
-                "name":"${referenceObject.name}",
-                "description":"${referenceObject.description}",
-                "config_type":"smtp_account",
-                "is_enabled":${referenceObject.isEnabled},
-                "smtp_account":$sampleSmtpJsonString
-            }
-        }
-        """.trimIndent()
-
-        val getUser = "getUser"
-        val getUserClient = SecureRestClientBuilder(clusterHosts.toTypedArray(), isHttps(), getUser, password)
-            .setSocketTimeout(60000)
-            .setConnectionRequestTimeout(180000)
-            .build()
-
-        try {
-            val configId = createConfigWithRequestJsonString(createRequestJsonString, userClient!!)
-            Assert.assertNotNull(configId)
-            Thread.sleep(1000)
-
-            createUserWithCustomRole(getUser, password, NOTIFICATION_GET_CONFIG_ACCESS, arrayOf("role2", "role1", "role3"), ROLE_TO_PERMISSION_MAPPING[NOTIFICATION_GET_CONFIG_ACCESS])
-
-            // Get SMTP account config
-            val getConfigResponse = executeRequest(
-                RestRequest.Method.GET.name,
-                "${NotificationPlugin.PLUGIN_BASE_URI}/configs/$configId",
-                "",
-                RestStatus.OK.status,
-                getUserClient
-            )
-        } finally {
-            deleteUserWithCustomRole(user, NOTIFICATION_CREATE_CONFIG_ACCESS)
-            deleteUserWithCustomRole(getUser, NOTIFICATION_GET_CONFIG_ACCESS)
-            getUserClient?.close()
-        }
-    }
-
-    fun `test update smtp sender has access when filter by backend access strategy is all`() {
-        updateClusterSettings(ClusterSetting("persistent", PluginSettings.FILTER_BY_BACKEND_ROLES.key, true))
-        updateClusterSettings(ClusterSetting("persistent", PluginSettings.FILTER_BY_BACKEND_ROLES_ACCESS_STRATEGY.key, FilterByBackendRolesAccessStrategy.ALL.strategy))
-
-        createUserWithCustomRole(user, password, NOTIFICATION_CREATE_CONFIG_ACCESS, arrayOf("role1"), ROLE_TO_PERMISSION_MAPPING[NOTIFICATION_CREATE_CONFIG_ACCESS])
-        // Create sample config request reference
-        val sampleSmtpAccount = SmtpAccount("example-host", 2465, MethodType.SSL, "no-reply@fake-host.com")
-        val referenceObject = NotificationConfig(
-            "this is a sample config name",
-            "this is a sample config description",
-            ConfigType.SMTP_ACCOUNT,
-            isEnabled = true,
-            configData = sampleSmtpAccount
-        )
-        val sampleSmtpJsonString = getJsonString(sampleSmtpAccount)
-
-        // Create SMTP account config
-        val createRequestJsonString = """
-        {
-            "config":{
-                "name":"${referenceObject.name}",
-                "description":"${referenceObject.description}",
-                "config_type":"smtp_account",
-                "is_enabled":${referenceObject.isEnabled},
-                "smtp_account":$sampleSmtpJsonString
-            }
-        }
-        """.trimIndent()
-
-        // verifySingleConfigEquals(configId, referenceObject, getConfigResponse)
-
-        val updateUser = "updateUser"
-        val updateUserClient = SecureRestClientBuilder(clusterHosts.toTypedArray(), isHttps(), updateUser, password)
-            .setSocketTimeout(60000)
-            .setConnectionRequestTimeout(180000)
-            .build()
-
-        try {
-            val configId = createConfigWithRequestJsonString(createRequestJsonString, userClient!!)
-            Assert.assertNotNull(configId)
-            Thread.sleep(1000)
-
-            // roles on update user contain all roles from create user
-            createUserWithCustomRole(updateUser, password, NOTIFICATION_UPDATE_CONFIG_ACCESS, arrayOf("role1", "role2"), ROLE_TO_PERMISSION_MAPPING[NOTIFICATION_UPDATE_CONFIG_ACCESS])
-
-            val referenceObjectUpdate = NotificationConfig(
-                "this is a sample config name updated",
-                "this is a sample config description updated",
-                ConfigType.SMTP_ACCOUNT,
-                isEnabled = true,
-                configData = sampleSmtpAccount
-            )
-            val updateRequestJsonString = """
-            {
-                "config":{
-                    "name":"${referenceObjectUpdate.name}",
-                    "description":"${referenceObjectUpdate.description}",
-                    "config_type":"smtp_account",
-                    "is_enabled":${referenceObject.isEnabled},
-                    "smtp_account":$sampleSmtpJsonString
-                }
-            }
-            """.trimIndent()
-            executeRequest(
-                RestRequest.Method.PUT.name,
-                "${NotificationPlugin.PLUGIN_BASE_URI}/configs/$configId",
-                updateRequestJsonString,
-                RestStatus.OK.status,
-                updateUserClient
-            )
-            Thread.sleep(1000)
-
-            // Get SMTP account config
-            val getConfigResponse = executeRequest(
-                RestRequest.Method.GET.name,
-                "${NotificationPlugin.PLUGIN_BASE_URI}/configs/$configId",
-                "",
-                RestStatus.OK.status
-            )
-            verifySingleConfigEquals(configId, referenceObjectUpdate, getConfigResponse)
-        } finally {
-            deleteUserWithCustomRole(user, NOTIFICATION_CREATE_CONFIG_ACCESS)
-            deleteUserWithCustomRole(updateUser, NOTIFICATION_UPDATE_CONFIG_ACCESS)
-            updateUserClient?.close()
-        }
-    }
-
-    fun `test update smtp sender does not have access when filter by backend access strategy is all`() {
-        updateClusterSettings(ClusterSetting("persistent", PluginSettings.FILTER_BY_BACKEND_ROLES.key, true))
-        updateClusterSettings(ClusterSetting("persistent", PluginSettings.FILTER_BY_BACKEND_ROLES_ACCESS_STRATEGY.key, FilterByBackendRolesAccessStrategy.ALL.strategy))
-
-        createUserWithCustomRole(user, password, NOTIFICATION_CREATE_CONFIG_ACCESS, arrayOf("role1", "role2"), ROLE_TO_PERMISSION_MAPPING[NOTIFICATION_CREATE_CONFIG_ACCESS])
-        // Create sample config request reference
-        val sampleSmtpAccount = SmtpAccount("example-host", 2465, MethodType.SSL, "no-reply@fake-host.com")
-        val referenceObject = NotificationConfig(
-            "this is a sample config name",
-            "this is a sample config description",
-            ConfigType.SMTP_ACCOUNT,
-            isEnabled = true,
-            configData = sampleSmtpAccount
-        )
-        val sampleSmtpJsonString = getJsonString(sampleSmtpAccount)
-
-        // Create SMTP account config
-        val createRequestJsonString = """
-        {
-            "config":{
-                "name":"${referenceObject.name}",
-                "description":"${referenceObject.description}",
-                "config_type":"smtp_account",
-                "is_enabled":${referenceObject.isEnabled},
-                "smtp_account":$sampleSmtpJsonString
-            }
-        }
-        """.trimIndent()
-
-        // verifySingleConfigEquals(configId, referenceObject, getConfigResponse)
-
-        val updateUser = "updateUser"
-        val updateUserClient = SecureRestClientBuilder(clusterHosts.toTypedArray(), isHttps(), updateUser, password)
-            .setSocketTimeout(60000)
-            .setConnectionRequestTimeout(180000)
-            .build()
-
-        try {
-            val configId = createConfigWithRequestJsonString(createRequestJsonString, userClient!!)
-            Assert.assertNotNull(configId)
-            Thread.sleep(1000)
-
-            // roles on update user does not contain all roles from create user
-            createUserWithCustomRole(updateUser, password, NOTIFICATION_UPDATE_CONFIG_ACCESS, arrayOf("role1"), ROLE_TO_PERMISSION_MAPPING[NOTIFICATION_UPDATE_CONFIG_ACCESS])
-
-            val referenceObjectUpdate = NotificationConfig(
-                "this is a sample config name updated",
-                "this is a sample config description updated",
-                ConfigType.SMTP_ACCOUNT,
-                isEnabled = true,
-                configData = sampleSmtpAccount
-            )
-            val updateRequestJsonString = """
-            {
-                "config":{
-                    "name":"${referenceObjectUpdate.name}",
-                    "description":"${referenceObjectUpdate.description}",
-                    "config_type":"smtp_account",
-                    "is_enabled":${referenceObject.isEnabled},
-                    "smtp_account":$sampleSmtpJsonString
-                }
-            }
-            """.trimIndent()
-            executeRequest(
-                RestRequest.Method.PUT.name,
-                "${NotificationPlugin.PLUGIN_BASE_URI}/configs/$configId",
-                updateRequestJsonString,
-                RestStatus.FORBIDDEN.status,
-                updateUserClient
-            )
-        } finally {
-            deleteUserWithCustomRole(user, NOTIFICATION_CREATE_CONFIG_ACCESS)
-            deleteUserWithCustomRole(updateUser, NOTIFICATION_UPDATE_CONFIG_ACCESS)
-            updateUserClient?.close()
-        }
-    }
-
-    fun `test update smtp sender has access when filter by backend access strategy is exact`() {
-        updateClusterSettings(ClusterSetting("persistent", PluginSettings.FILTER_BY_BACKEND_ROLES.key, true))
-        updateClusterSettings(ClusterSetting("persistent", PluginSettings.FILTER_BY_BACKEND_ROLES_ACCESS_STRATEGY.key, FilterByBackendRolesAccessStrategy.EXACT.strategy))
-
-        createUserWithCustomRole(user, password, NOTIFICATION_CREATE_CONFIG_ACCESS, arrayOf("role1", "role2"), ROLE_TO_PERMISSION_MAPPING[NOTIFICATION_CREATE_CONFIG_ACCESS])
-        // Create sample config request reference
-        val sampleSmtpAccount = SmtpAccount("example-host", 2465, MethodType.SSL, "no-reply@fake-host.com")
-        val referenceObject = NotificationConfig(
-            "this is a sample config name",
-            "this is a sample config description",
-            ConfigType.SMTP_ACCOUNT,
-            isEnabled = true,
-            configData = sampleSmtpAccount
-        )
-        val sampleSmtpJsonString = getJsonString(sampleSmtpAccount)
-
-        // Create SMTP account config
-        val createRequestJsonString = """
-        {
-            "config":{
-                "name":"${referenceObject.name}",
-                "description":"${referenceObject.description}",
-                "config_type":"smtp_account",
-                "is_enabled":${referenceObject.isEnabled},
-                "smtp_account":$sampleSmtpJsonString
-            }
-        }
-        """.trimIndent()
-
-        val updateUser = "updateUser"
-        val updateUserClient = SecureRestClientBuilder(clusterHosts.toTypedArray(), isHttps(), updateUser, password)
-            .setSocketTimeout(60000)
-            .setConnectionRequestTimeout(180000)
-            .build()
-
-        try {
-            val configId = createConfigWithRequestJsonString(createRequestJsonString, userClient!!)
-            Assert.assertNotNull(configId)
-            Thread.sleep(1000)
-
-            // roles on update user contain all roles from create user
-            createUserWithCustomRole(updateUser, password, NOTIFICATION_UPDATE_CONFIG_ACCESS, arrayOf("role1", "role2"), ROLE_TO_PERMISSION_MAPPING[NOTIFICATION_UPDATE_CONFIG_ACCESS])
-
-            val referenceObjectUpdate = NotificationConfig(
-                "this is a sample config name updated",
-                "this is a sample config description updated",
-                ConfigType.SMTP_ACCOUNT,
-                isEnabled = true,
-                configData = sampleSmtpAccount
-            )
-            val updateRequestJsonString = """
-            {
-                "config":{
-                    "name":"${referenceObjectUpdate.name}",
-                    "description":"${referenceObjectUpdate.description}",
-                    "config_type":"smtp_account",
-                    "is_enabled":${referenceObject.isEnabled},
-                    "smtp_account":$sampleSmtpJsonString
-                }
-            }
-            """.trimIndent()
-            executeRequest(
-                RestRequest.Method.PUT.name,
-                "${NotificationPlugin.PLUGIN_BASE_URI}/configs/$configId",
-                updateRequestJsonString,
-                RestStatus.OK.status,
-                updateUserClient
-            )
-            Thread.sleep(1000)
-
-            // Get SMTP account config
-            val getConfigResponse = executeRequest(
-                RestRequest.Method.GET.name,
-                "${NotificationPlugin.PLUGIN_BASE_URI}/configs/$configId",
-                "",
-                RestStatus.OK.status
-            )
-            verifySingleConfigEquals(configId, referenceObjectUpdate, getConfigResponse)
-        } finally {
-            deleteUserWithCustomRole(user, NOTIFICATION_CREATE_CONFIG_ACCESS)
-            deleteUserWithCustomRole(updateUser, NOTIFICATION_UPDATE_CONFIG_ACCESS)
-            updateUserClient?.close()
-        }
-    }
-
-    fun `test update smtp sender does not have access when filter by backend access strategy is exact`() {
-        updateClusterSettings(ClusterSetting("persistent", PluginSettings.FILTER_BY_BACKEND_ROLES.key, true))
-        updateClusterSettings(ClusterSetting("persistent", PluginSettings.FILTER_BY_BACKEND_ROLES_ACCESS_STRATEGY.key, FilterByBackendRolesAccessStrategy.ALL.strategy))
-
-        createUserWithCustomRole(user, password, NOTIFICATION_CREATE_CONFIG_ACCESS, arrayOf("role1", "role2"), ROLE_TO_PERMISSION_MAPPING[NOTIFICATION_CREATE_CONFIG_ACCESS])
-        // Create sample config request reference
-        val sampleSmtpAccount = SmtpAccount("example-host", 2465, MethodType.SSL, "no-reply@fake-host.com")
-        val referenceObject = NotificationConfig(
-            "this is a sample config name",
-            "this is a sample config description",
-            ConfigType.SMTP_ACCOUNT,
-            isEnabled = true,
-            configData = sampleSmtpAccount
-        )
-        val sampleSmtpJsonString = getJsonString(sampleSmtpAccount)
-
-        // Create SMTP account config
-        val createRequestJsonString = """
-        {
-            "config":{
-                "name":"${referenceObject.name}",
-                "description":"${referenceObject.description}",
-                "config_type":"smtp_account",
-                "is_enabled":${referenceObject.isEnabled},
-                "smtp_account":$sampleSmtpJsonString
-            }
-        }
-        """.trimIndent()
-
-        val updateUser = "updateUser"
-        val updateUserClient = SecureRestClientBuilder(clusterHosts.toTypedArray(), isHttps(), updateUser, password)
-            .setSocketTimeout(60000)
-            .setConnectionRequestTimeout(180000)
-            .build()
-
-        try {
-            val configId = createConfigWithRequestJsonString(createRequestJsonString, userClient!!)
-            Assert.assertNotNull(configId)
-            Thread.sleep(1000)
-
-            // roles on update user does match all roles from create user
-            createUserWithCustomRole(updateUser, password, NOTIFICATION_UPDATE_CONFIG_ACCESS, arrayOf("role1"), ROLE_TO_PERMISSION_MAPPING[NOTIFICATION_UPDATE_CONFIG_ACCESS])
-
-            val referenceObjectUpdate = NotificationConfig(
-                "this is a sample config name updated",
-                "this is a sample config description updated",
-                ConfigType.SMTP_ACCOUNT,
-                isEnabled = true,
-                configData = sampleSmtpAccount
-            )
-            val updateRequestJsonString = """
-            {
-                "config":{
-                    "name":"${referenceObjectUpdate.name}",
-                    "description":"${referenceObjectUpdate.description}",
-                    "config_type":"smtp_account",
-                    "is_enabled":${referenceObject.isEnabled},
-                    "smtp_account":$sampleSmtpJsonString
-                }
-            }
-            """.trimIndent()
-            executeRequest(
-                RestRequest.Method.PUT.name,
-                "${NotificationPlugin.PLUGIN_BASE_URI}/configs/$configId",
-                updateRequestJsonString,
-                RestStatus.FORBIDDEN.status,
-                updateUserClient
-            )
-        } finally {
-            deleteUserWithCustomRole(user, NOTIFICATION_CREATE_CONFIG_ACCESS)
-            deleteUserWithCustomRole(updateUser, NOTIFICATION_UPDATE_CONFIG_ACCESS)
-            updateUserClient?.close()
-        }
-    }
-
-    fun `test delete smtp sender has access when filter by backend access strategy is all`() {
-        updateClusterSettings(ClusterSetting("persistent", PluginSettings.FILTER_BY_BACKEND_ROLES.key, true))
-        updateClusterSettings(ClusterSetting("persistent", PluginSettings.FILTER_BY_BACKEND_ROLES_ACCESS_STRATEGY.key, FilterByBackendRolesAccessStrategy.ALL.strategy))
-
-        createUserWithCustomRole(user, password, NOTIFICATION_CREATE_CONFIG_ACCESS, arrayOf("role1", "role2"), ROLE_TO_PERMISSION_MAPPING[NOTIFICATION_CREATE_CONFIG_ACCESS])
-        // Create sample config request reference
-        val sampleSmtpAccount = SmtpAccount("example-host", 2465, MethodType.SSL, "no-reply@fake-host.com")
-        val referenceObject = NotificationConfig(
-            "this is a sample config name",
-            "this is a sample config description",
-            ConfigType.SMTP_ACCOUNT,
-            isEnabled = true,
-            configData = sampleSmtpAccount
-        )
-        val sampleSmtpJsonString = getJsonString(sampleSmtpAccount)
-
-        // Create SMTP account config
-        val createRequestJsonString = """
-        {
-            "config":{
-                "name":"${referenceObject.name}",
-                "description":"${referenceObject.description}",
-                "config_type":"smtp_account",
-                "is_enabled":${referenceObject.isEnabled},
-                "smtp_account":$sampleSmtpJsonString
-            }
-        }
-        """.trimIndent()
-
-        val deleteUser = "deleteUser"
-        val deleteUserClient = SecureRestClientBuilder(clusterHosts.toTypedArray(), isHttps(), deleteUser, password)
-            .setSocketTimeout(60000)
-            .setConnectionRequestTimeout(180000)
-            .build()
-
-        try {
-            val configId = createConfigWithRequestJsonString(createRequestJsonString, userClient!!)
-            Assert.assertNotNull(configId)
-            Thread.sleep(1000)
-
-            // roles on update user contain all roles from create user
-            createUserWithCustomRole(deleteUser, password, NOTIFICATION_DELETE_CONFIG_ACCESS, arrayOf("role1", "role2"), ROLE_TO_PERMISSION_MAPPING[NOTIFICATION_DELETE_CONFIG_ACCESS])
-
-            // Get SMTP account config
-            val getConfigResponse = executeRequest(
-                RestRequest.Method.DELETE.name,
-                "${NotificationPlugin.PLUGIN_BASE_URI}/configs/$configId",
-                "",
-                RestStatus.OK.status,
-                deleteUserClient
-            )
-        } finally {
-            deleteUserWithCustomRole(user, NOTIFICATION_CREATE_CONFIG_ACCESS)
-            deleteUserWithCustomRole(deleteUser, NOTIFICATION_DELETE_CONFIG_ACCESS)
-            deleteUserClient?.close()
-        }
-    }
-
-    fun `test delete smtp sender does not have access when filter by backend access strategy is all`() {
-        updateClusterSettings(ClusterSetting("persistent", PluginSettings.FILTER_BY_BACKEND_ROLES.key, true))
-        updateClusterSettings(ClusterSetting("persistent", PluginSettings.FILTER_BY_BACKEND_ROLES_ACCESS_STRATEGY.key, FilterByBackendRolesAccessStrategy.ALL.strategy))
-
-        createUserWithCustomRole(user, password, NOTIFICATION_CREATE_CONFIG_ACCESS, arrayOf("role1", "role2"), ROLE_TO_PERMISSION_MAPPING[NOTIFICATION_CREATE_CONFIG_ACCESS])
-        // Create sample config request reference
-        val sampleSmtpAccount = SmtpAccount("example-host", 2465, MethodType.SSL, "no-reply@fake-host.com")
-        val referenceObject = NotificationConfig(
-            "this is a sample config name",
-            "this is a sample config description",
-            ConfigType.SMTP_ACCOUNT,
-            isEnabled = true,
-            configData = sampleSmtpAccount
-        )
-        val sampleSmtpJsonString = getJsonString(sampleSmtpAccount)
-
-        // Create SMTP account config
-        val createRequestJsonString = """
-        {
-            "config":{
-                "name":"${referenceObject.name}",
-                "description":"${referenceObject.description}",
-                "config_type":"smtp_account",
-                "is_enabled":${referenceObject.isEnabled},
-                "smtp_account":$sampleSmtpJsonString
-            }
-        }
-        """.trimIndent()
-
-        val deleteUser = "deleteUser"
-        val deleteUserClient = SecureRestClientBuilder(clusterHosts.toTypedArray(), isHttps(), deleteUser, password)
-            .setSocketTimeout(60000)
-            .setConnectionRequestTimeout(180000)
-            .build()
-
-        try {
-            val configId = createConfigWithRequestJsonString(createRequestJsonString, userClient!!)
-            Assert.assertNotNull(configId)
-            Thread.sleep(1000)
-
-            // roles on update user does not contain all roles from create user
-            createUserWithCustomRole(deleteUser, password, NOTIFICATION_DELETE_CONFIG_ACCESS, arrayOf("role1"), ROLE_TO_PERMISSION_MAPPING[NOTIFICATION_DELETE_CONFIG_ACCESS])
-
-            executeRequest(
-                RestRequest.Method.DELETE.name,
-                "${NotificationPlugin.PLUGIN_BASE_URI}/configs/$configId",
-                "",
-                RestStatus.FORBIDDEN.status,
-                deleteUserClient
-            )
-        } finally {
-            deleteUserWithCustomRole(user, NOTIFICATION_CREATE_CONFIG_ACCESS)
-            deleteUserWithCustomRole(deleteUser, NOTIFICATION_DELETE_CONFIG_ACCESS)
-            deleteUserClient?.close()
-        }
-    }
-
-    fun `test delete smtp sender has access when filter by backend access strategy is exact`() {
-        updateClusterSettings(ClusterSetting("persistent", PluginSettings.FILTER_BY_BACKEND_ROLES.key, true))
-        updateClusterSettings(ClusterSetting("persistent", PluginSettings.FILTER_BY_BACKEND_ROLES_ACCESS_STRATEGY.key, FilterByBackendRolesAccessStrategy.EXACT.strategy))
-
-        createUserWithCustomRole(user, password, NOTIFICATION_CREATE_CONFIG_ACCESS, arrayOf("role1", "role2"), ROLE_TO_PERMISSION_MAPPING[NOTIFICATION_CREATE_CONFIG_ACCESS])
-        // Create sample config request reference
-        val sampleSmtpAccount = SmtpAccount("example-host", 2465, MethodType.SSL, "no-reply@fake-host.com")
-        val referenceObject = NotificationConfig(
-            "this is a sample config name",
-            "this is a sample config description",
-            ConfigType.SMTP_ACCOUNT,
-            isEnabled = true,
-            configData = sampleSmtpAccount
-        )
-        val sampleSmtpJsonString = getJsonString(sampleSmtpAccount)
-
-        // Create SMTP account config
-        val createRequestJsonString = """
-        {
-            "config":{
-                "name":"${referenceObject.name}",
-                "description":"${referenceObject.description}",
-                "config_type":"smtp_account",
-                "is_enabled":${referenceObject.isEnabled},
-                "smtp_account":$sampleSmtpJsonString
-            }
-        }
-        """.trimIndent()
-
-        val deleteUser = "deleteUser"
-        val deleteUserClient = SecureRestClientBuilder(clusterHosts.toTypedArray(), isHttps(), deleteUser, password)
-            .setSocketTimeout(60000)
-            .setConnectionRequestTimeout(180000)
-            .build()
-
-        try {
-            val configId = createConfigWithRequestJsonString(createRequestJsonString, userClient!!)
-            Assert.assertNotNull(configId)
-            Thread.sleep(1000)
-
-            // roles on update user matches all roles from create user
-            createUserWithCustomRole(deleteUser, password, NOTIFICATION_DELETE_CONFIG_ACCESS, arrayOf("role1", "role2"), ROLE_TO_PERMISSION_MAPPING[NOTIFICATION_DELETE_CONFIG_ACCESS])
-
-            // Get SMTP account config
-            val getConfigResponse = executeRequest(
-                RestRequest.Method.DELETE.name,
-                "${NotificationPlugin.PLUGIN_BASE_URI}/configs/$configId",
-                "",
-                RestStatus.OK.status,
-                deleteUserClient
-            )
-        } finally {
-            deleteUserWithCustomRole(user, NOTIFICATION_CREATE_CONFIG_ACCESS)
-            deleteUserWithCustomRole(deleteUser, NOTIFICATION_DELETE_CONFIG_ACCESS)
-            deleteUserClient?.close()
-        }
-    }
-
-    fun `test delete smtp sender does not have access when filter by backend access strategy is exact`() {
-        updateClusterSettings(ClusterSetting("persistent", PluginSettings.FILTER_BY_BACKEND_ROLES.key, true))
-        updateClusterSettings(ClusterSetting("persistent", PluginSettings.FILTER_BY_BACKEND_ROLES_ACCESS_STRATEGY.key, FilterByBackendRolesAccessStrategy.EXACT.strategy))
-
-        createUserWithCustomRole(user, password, NOTIFICATION_CREATE_CONFIG_ACCESS, arrayOf("role1", "role2"), ROLE_TO_PERMISSION_MAPPING[NOTIFICATION_CREATE_CONFIG_ACCESS])
-        // Create sample config request reference
-        val sampleSmtpAccount = SmtpAccount("example-host", 2465, MethodType.SSL, "no-reply@fake-host.com")
-        val referenceObject = NotificationConfig(
-            "this is a sample config name",
-            "this is a sample config description",
-            ConfigType.SMTP_ACCOUNT,
-            isEnabled = true,
-            configData = sampleSmtpAccount
-        )
-        val sampleSmtpJsonString = getJsonString(sampleSmtpAccount)
-
-        // Create SMTP account config
-        val createRequestJsonString = """
-        {
-            "config":{
-                "name":"${referenceObject.name}",
-                "description":"${referenceObject.description}",
-                "config_type":"smtp_account",
-                "is_enabled":${referenceObject.isEnabled},
-                "smtp_account":$sampleSmtpJsonString
-            }
-        }
-        """.trimIndent()
-
-        val deleteUser = "deleteUser"
-        val deleteUserClient = SecureRestClientBuilder(clusterHosts.toTypedArray(), isHttps(), deleteUser, password)
-            .setSocketTimeout(60000)
-            .setConnectionRequestTimeout(180000)
-            .build()
-
-        try {
-            val configId = createConfigWithRequestJsonString(createRequestJsonString, userClient!!)
-            Assert.assertNotNull(configId)
-            Thread.sleep(1000)
-
-            // roles on update user does not match all roles from create user
-            createUserWithCustomRole(deleteUser, password, NOTIFICATION_DELETE_CONFIG_ACCESS, arrayOf("role1"), ROLE_TO_PERMISSION_MAPPING[NOTIFICATION_DELETE_CONFIG_ACCESS])
-
-            executeRequest(
-                RestRequest.Method.DELETE.name,
-                "${NotificationPlugin.PLUGIN_BASE_URI}/configs/$configId",
-                "",
-                RestStatus.FORBIDDEN.status,
-                deleteUserClient
-            )
-        } finally {
-            deleteUserWithCustomRole(user, NOTIFICATION_CREATE_CONFIG_ACCESS)
-            deleteUserWithCustomRole(deleteUser, NOTIFICATION_DELETE_CONFIG_ACCESS)
-            deleteUserClient?.close()
-        }
-    }
 }
